+++ platform/ui/src/components/ViewportActionCorners/index.tsx
@@ -1,2 +1,4 @@
-import ViewportActionCorners from './ViewportActionCorners';
+import ViewportActionCorners, { ViewportActionCornersLocations } from './ViewportActionCorners';
+
+export { ViewportActionCornersLocations };
 export default ViewportActionCorners;
