+++ platform/ui/src/components/index.js
@@ -73,7 +73,7 @@ import CheckBox from './CheckBox';
 import LoadingIndicatorProgress from './LoadingIndicatorProgress';
 import LoadingIndicatorTotalPercent from './LoadingIndicatorTotalPercent';
 import ViewportActionBar from './ViewportActionBar';
-import ViewportActionCorners from './ViewportActionCorners';
+import ViewportActionCorners, { ViewportActionCornersLocations } from './ViewportActionCorners';
 import ProgressLoadingBar from './ProgressLoadingBar';
 import LegacySidePanel from './LegacySidePanel';
 import PanelSection from './PanelSection';
@@ -167,6 +167,7 @@ export {
   ViewportActionArrows,
   ViewportActionBar,
   ViewportActionCorners,
+  ViewportActionCornersLocations,
   ViewportDownloadForm,
   ViewportGrid,
   ViewportPane,
