+++ platform/ui/src/types/ViewportActionCornersTypes.ts
@@ -1,20 +1,6 @@
 import { ReactNode } from 'react';
 
-export enum ViewportActionCornersLocations {
-  topLeft,
-  topRight,
-  bottomLeft,
-  bottomRight,
-}
-
 export interface ViewportActionCornersComponentInfo {
   id: string;
   component: ReactNode;
 }
-
-export type ViewportActionCornersProps = {
-  cornerComponents: Record<
-    ViewportActionCornersLocations,
-    Array<ViewportActionCornersComponentInfo>
-  >;
-};
